+++ src/main/java/com/leonardobishop/playerskills2/skills/Skill.java
@@ -75,7 +75,7 @@ public abstract class Skill implements Listener {
         }
 
         if (obj instanceof Integer) {
-            return ((double) ((int) obj));
+            return ((int) obj);
         } else {
             return (double) obj;
         }
